Preliminary LV2 UI support
LV2 support is enabled by using a subset of the full DSSI GUI. Rather than
create top-level windows for the patch list and edit window, just
the core editing control (the notepad global variable in gui_interface.c) is
instantiated and passed to the LV2 host. Any controls found in other windows or
containers are not accessible; notably tuning, monophonic and glide controls are
not accessible.

OSC communication is replaced by calls to lv2_write_function when sending events
to the host, and the port_event function to receive events from the host.

TODO: thread safety