LV2 UI thread safety
The big change here is to use dynamic allocation for the voice_widgets array.
This array is now allocated by the create_edit_window function, with pointers to
it stored in an array of y_callback_data_t structs.

The y_callback_data_t type is used to store the information required by the
various UI callback functions; primarily a pointer to the voice_widgets array
and, when running as a LV2 plugin, a pointer the LV2_Write_function. An array of
y_callback_data_t is created on startup, with pointers to its elements passed
into the various UI callback functions.

The changes are fairly simple, but there's many required due to the nature of
having to pass around pointers, sometimes through several functions, to formerly
global variables.