Prevent UI freezing when the LV2 UI is loaded multiple times.
When using the macros at the top of gtkknob.h, the GtkKnob component will be
registered via the gobject type system. To prevent multiple registration of the
same type, a static variable in the gtk_knob_get_type() function (implemented
via macros in gobject/gtype.h in the glib library) indicates if the type has
been registered.

This static variable is reset whenever the UI is loaded via dlopen(). When this
occurs, calls to the gtk_knob_get_type() function will attempt to re-register the
GtkKnob. This registration fails, preventing the call to g_once_init_leave()
in gtk_knob_get_type() from succeeding.

The next call to gtk_knob_get_type() will then block in the g_once_init_enter()
function, waiting for a call to g_once_init_leave(), freezing the UI.

This workaround checks for this condition and retrieves the type id via the existing
registration, passing it into g_once_init_leave(), allowing it to succeed.

Since GtkKnob is not registered on subsequent loads, gtk_knob_class_init() is
not called. Instead, we assign the static parent_class variable (in gtkknob.c)
to enable gtk_knob_destroy() and gtk_knob_finalize() to function correctly.